requirements.txt
aiohttp==3.6.2
async-timeout==3.0.1
attrs==19.3.0
cachetools==4.1.1
certifi==2020.6.20
cffi==1.14.0
chardet==3.0.4
discord.py==1.3.3
dnspython==1.16.0
google-api-core==1.21.0
google-api-python-client==1.9.3
google-auth==1.18.0
google-auth-httplib2==0.0.3
googleapis-common-protos==1.52.0
httplib2==0.18.1
idna==2.10
idna-ssl==1.1.0
multidict==4.7.6
protobuf==3.12.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pymongo==3.10.1
PyNaCl==1.3.0
pytz==2020.1
requests==2.24.0
rsa==4.6
six==1.15.0
typing-extensions==3.7.4.2
uritemplate==3.0.1
urllib3==1.25.9
websockets==8.1
yarl==1.4.2
